chore(artifacts): split gcs reference tests to its own file
Fixed the version check...
