+++ tests/system_tests/test_artifacts/test_references_gcs.py
@@ -1,0 +1,232 @@
+from __future__ import annotations
+
+import os
+
+from pytest import fixture, raises
+from wandb import Artifact
+from wandb.sdk.artifacts.artifact_manifest_entry import ArtifactManifestEntry
+from wandb.sdk.artifacts.artifact_state import ArtifactState
+from wandb.sdk.artifacts.storage_handlers.gcs_handler import (
+    GCSHandler,
+    _GCSIsADirectoryError,
+)
+
+
+@fixture
+def artifact() -> Artifact:
+    return Artifact(type="dataset", name="data-artifact")
+
+
+def mock_gcs(artifact, override_blob_name="my_object.pb", path=False, hash=True):
+    class Blob:
+        def __init__(
+            self, name: str = override_blob_name, metadata=None, generation: int = 1
+        ):
+            self.md5_hash = "1234567890abcde" if hash else None
+            self.etag = "1234567890abcde"
+            self.generation = generation
+            self.name = name
+            self.size = 10
+
+    class GSBucket:
+        def __init__(self):
+            self.versioning_enabled = True
+
+        def reload(self, *args, **kwargs):
+            return
+
+        def get_blob(self, key=override_blob_name, *args, **kwargs):
+            return (
+                None
+                if path or key != override_blob_name
+                else Blob(generation=kwargs.get("generation"))
+            )
+
+        # https://docs.cloud.google.com/python/docs/reference/storage/latest/google.cloud.storage.bucket.Bucket#google_cloud_storage_bucket_Bucket_list_blobs
+        def list_blobs(self, *args, **kwargs):
+            versions: bool = kwargs.get("versions", False)
+            prefix: str = kwargs.get("prefix", "")
+            # For versioned lookups, return blobs with different generations
+            if versions and prefix == override_blob_name:
+                return [
+                    Blob(name=override_blob_name, generation=1),
+                    Blob(name=override_blob_name, generation=2),
+                    Blob(name=override_blob_name, generation=3),
+                ]
+            # For directory paths (ends with /)
+            if override_blob_name.endswith("/"):
+                return [
+                    Blob(name=override_blob_name),
+                    Blob(name=os.path.join(override_blob_name, "my_other_object.pb")),
+                ]
+            # For single file lookup (prefix matches exact filename)
+            if prefix == override_blob_name and not path:
+                return [Blob(name=override_blob_name)]
+            # For directory listing (path=True means treat as directory)
+            if path or prefix == "":
+                return [
+                    Blob(name="my_object.pb"),
+                    Blob(name="my_other_object.pb"),
+                ]
+            # Default: return just the matching blob
+            return [Blob(name=override_blob_name)]
+
+    class GSClient:
+        def bucket(self, bucket):
+            return GSBucket()
+
+    mock = GSClient()
+    for handler in artifact.manifest.storage_policy._handler._handlers:
+        if isinstance(handler, GCSHandler):
+            handler._client = mock
+    return mock
+
+
+def test_add_gs_reference_object(artifact):
+    mock_gcs(artifact)
+    artifact.add_reference("gs://my-bucket/my_object.pb")
+
+    assert artifact.digest == "8aec0d6978da8c2b0bf5662b3fd043a4"
+    manifest_contents = artifact.manifest.to_manifest_json()["contents"]
+    assert manifest_contents == {
+        "my_object.pb": {
+            "digest": "1234567890abcde",
+            "ref": "gs://my-bucket/my_object.pb",
+            "extra": {"versionID": 1},
+            "size": 10,
+        },
+    }
+
+
+def test_load_gs_reference_object_without_generation_and_mismatched_etag(
+    artifact,
+):
+    mock_gcs(artifact)
+    artifact.add_reference("gs://my-bucket/my_object.pb")
+    artifact._state = ArtifactState.COMMITTED
+    entry = artifact.get_entry("my_object.pb")
+    entry.extra = {}
+    entry.digest = "abad0"
+
+    with raises(ValueError, match="Digest mismatch"):
+        entry.download()
+
+
+def test_add_gs_reference_object_with_version(artifact):
+    mock_gcs(artifact)
+    artifact.add_reference("gs://my-bucket/my_object.pb#2")
+
+    assert artifact.digest == "8aec0d6978da8c2b0bf5662b3fd043a4"
+    manifest_contents = artifact.manifest.to_manifest_json()["contents"]
+    assert manifest_contents == {
+        "my_object.pb": {
+            "digest": "1234567890abcde",
+            "ref": "gs://my-bucket/my_object.pb",
+            "extra": {"versionID": 2},
+            "size": 10,
+        },
+    }
+
+
+def test_add_gs_reference_object_with_name(artifact):
+    mock_gcs(artifact)
+    artifact.add_reference("gs://my-bucket/my_object.pb", name="renamed.pb")
+
+    assert artifact.digest == "bd85fe009dc9e408a5ed9b55c95f47b2"
+    manifest_contents = artifact.manifest.to_manifest_json()["contents"]
+    assert manifest_contents == {
+        "renamed.pb": {
+            "digest": "1234567890abcde",
+            "ref": "gs://my-bucket/my_object.pb",
+            "extra": {"versionID": 1},
+            "size": 10,
+        },
+    }
+
+
+def test_add_gs_reference_path(capsys, artifact):
+    mock_gcs(artifact, path=True)
+    artifact.add_reference("gs://my-bucket/")
+
+    assert artifact.digest == "17955d00a20e1074c3bc96c74b724bfe"
+    manifest_contents = artifact.manifest.to_manifest_json()["contents"]
+    assert manifest_contents == {
+        "my_object.pb": {
+            "digest": "1234567890abcde",
+            "ref": "gs://my-bucket/my_object.pb",
+            "extra": {"versionID": 1},
+            "size": 10,
+        },
+        "my_other_object.pb": {
+            "digest": "1234567890abcde",
+            "ref": "gs://my-bucket/my_other_object.pb",
+            "extra": {"versionID": 1},
+            "size": 10,
+        },
+    }
+    _, err = capsys.readouterr()
+    assert "Added 2 objects" in err
+
+
+def test_add_gs_reference_object_no_md5(artifact):
+    mock_gcs(artifact, hash=False)
+    artifact.add_reference("gs://my-bucket/my_object.pb")
+
+    assert artifact.digest == "8aec0d6978da8c2b0bf5662b3fd043a4"
+    manifest_contents = artifact.manifest.to_manifest_json()["contents"]
+    assert manifest_contents == {
+        "my_object.pb": {
+            "digest": "1234567890abcde",
+            "ref": "gs://my-bucket/my_object.pb",
+            "extra": {"versionID": 1},
+            "size": 10,
+        },
+    }
+
+
+def test_add_gs_reference_with_dir_paths(artifact):
+    mock_gcs(artifact, override_blob_name="my_folder/")
+    artifact.add_reference("gs://my-bucket/my_folder/")
+
+    # uploading a reference to a folder path should add entries for
+    # everything returned by the list_blobs call
+    assert len(artifact.manifest.entries) == 1
+    manifest_contents = artifact.manifest.to_manifest_json()["contents"]
+    assert manifest_contents == {
+        "my_other_object.pb": {
+            "digest": "1234567890abcde",
+            "ref": "gs://my-bucket/my_folder/my_other_object.pb",
+            "extra": {"versionID": 1},
+            "size": 10,
+        },
+    }
+
+
+def test_load_gs_reference_with_dir_paths(artifact):
+    mock = mock_gcs(artifact, override_blob_name="my_folder/")
+    artifact.add_reference("gs://my-bucket/my_folder/")
+
+    gcs_handler = GCSHandler()
+    gcs_handler._client = mock
+
+    # simple case where ref ends with "/"
+    simple_entry = ArtifactManifestEntry(
+        path="my-bucket/my_folder",
+        ref="gs://my-bucket/my_folder/",
+        digest="1234567890abcde",
+        size=0,
+        extra={"versionID": 1},
+    )
+    with raises(_GCSIsADirectoryError):
+        gcs_handler.load_path(simple_entry, local=True)
+
+    # case where we didn't store "/" and have to use get_blob
+    entry = ArtifactManifestEntry(
+        path="my-bucket/my_folder",
+        ref="gs://my-bucket/my_folder",
+        digest="1234567890abcde",
+        size=0,
+        extra={"versionID": 1},
+    )
+    with raises(_GCSIsADirectoryError):
+        gcs_handler.load_path(entry, local=True)
